Fix num to double type conversion error in modern_ui_components.dart

+++ lib/widget/modern_ui_components.dart
@@ -228,7 +228,7 @@ class ProgressIndicatorCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final percentage = (progress * 100).clamp(0, 100);
+    final percentage = (progress * 100).clamp(0, 100).toDouble();
     final statusColor = SFMSTheme.getStatusColor(percentage);
     final statusGradient = SFMSTheme.getStatusGradient(percentage);
